requirements.txt
# Core FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Machine Learning and AI
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
scikit-learn==1.3.2
numpy==1.24.3
scipy==1.11.4

# Image processing
Pillow==10.1.0

# Data processing
pandas==2.1.3
PyYAML==6.0.1

# Database and caching
asyncpg==0.29.0
aioredis==2.0.1
sqlalchemy[asyncio]==2.0.23

# AWS and cloud services
boto3==1.34.0
botocore==1.34.0

# Monitoring and observability
prometheus-client==0.19.0

# HTTP and networking
aiohttp==3.9.1
aiofiles==23.2.0
httpx==0.25.2

# System monitoring
psutil==5.9.6

# GPU monitoring (optional)
pynvml==11.5.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Logging and utilities
structlog==23.2.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# CORS and security
python-cors==1.7.0

# Environment and configuration
python-dotenv==1.0.0

